add missing params tests
